move BusinessPrimaryActionComponent to Businesses namespace
