+++ app/views/businesses/show.html.erb
@@ -24,7 +24,7 @@
       </div>
 
       <div class="mt-6 flex flex-col-reverse justify-stretch space-y-4 space-y-reverse sm:flex-row-reverse sm:justify-end sm:space-x-reverse sm:space-y-0 sm:space-x-3 md:mt-0 md:flex-row md:space-x-3">
-        <%= render BusinessPrimaryActionComponent.new(@business, user: current_user) %>
+        <%= render Businesses::PrimaryActionComponent.new(@business, user: current_user) %>
       </div>
     </div>
